requirements.txt

Flask==2.2.3
google-generativeai
gunicorn==20.0.4
grpcio

grpcio
Werkzeug==2.2.3
Flask
Flask-Cors
requests
python-dotenv
groq

flask
requests
fastapi
uvicorn
pydantic
mysql-connector-python

